+++ src/http/middleware.rs
@@ -1,2 +1,3 @@
+#![allow(dead_code)] // Not intended to be used yet
 pub mod external_token;
 pub mod extract_external_token;
