+++ src/http/mod.rs
@@ -1,6 +1,5 @@
 pub mod controllers;
 /// This module contains functions references HTTP-related entities such as requests, responses, and routes.
-mod conversions;
 pub mod errors;
 pub mod health;
 mod middleware;
